Improve performance of feature_store.get
nextTick appears to perform much faster than setTimeout and marginally faster than setImmediate in environments we've tested in.
(AWS linux)

This nodejs event loop blog post describes when to use setImmediate vs process.nextTick. Although it recommends setImmediate for most cases, in this case, process.nextTick feels more appropriate.
https://nodejs.org/en/docs/guides/event-loop-timers-and-nexttick/#process-nexttick-vs-setimmediate

The relative performance of process.nextTick, setImmediate and setTimeout can be dependent on the hardware, OS, nodeJS version according to this benchmark post. So we should test in a prod environment to ensure that replacing setTimeout with a more performant variant improves LD performance.

https://gist.github.com/mmalecki/1257394