Change some for ... in .. clauses to for ... of ...
Context: I am running a nodejs server that I was trying to integrate this library into.

However, our server uses some monkeypatched extensions of Arrays for convenience. Functions like `first`, `last`, `pluck`, etc (`Array.prototype.first = function() {...}`)

When trying to integrate this library into my project, it failed on the call to `init`, because this code was iterating not only over the indices in the arrays (0, 1, 2, etc), but also the names of our extension (first, last, pluck, etc).

This obviously threw an error when it tried to call `clause.op` on a Function.

I'm sure there's opinions out there either way on the practice of monkeypatching a global object type like Array, but regardless this made it much harder to integrate launchdarkly because I had to go through and remove all our extensions from the code.

As far as I know, `for ... of ...` is the recommended method to iterate over arrays as well.

I've gone ahead and changed all the instances of `for ... in ...` that were preventing library initialization. There may be more `for ... in ...` that should be converted to `for ... of ...`, but I suppose half the purpose of this PR is to draw your attention to this issue and get a conversation going.