Amend testdata ifMatch to use correct operator
The rule was not matching, after debugging it was found that the expected rule operator field
was named `op` not `operator`. This was found by debugging until getting to the matchFn function in
evaluator.js.

There was also another issue where updating when using an ifMatch resulted in a JSON ref error.  Replaced JSON based copy with manual copy: